+++ SharePointDsc/DSCResources/MSFT_SPSessionStateService/MSFT_SPSessionStateService.psm1
@@ -18,6 +18,14 @@ function Get-TargetResource
         $DatabaseServer,
 
         [Parameter()]
+        [System.Boolean]
+        $UseSQLAuthentication,
+
+        [Parameter()]
+        [System.Management.Automation.PSCredential]
+        $DatabaseCredentials,
+
+        [Parameter()]
         [ValidateSet("Present", "Absent")]
         [System.String]
         $Ensure = "Present",
@@ -66,6 +74,14 @@ function Set-TargetResource
         $DatabaseServer,
 
         [Parameter()]
+        [System.Boolean]
+        $UseSQLAuthentication,
+
+        [Parameter()]
+        [System.Management.Automation.PSCredential]
+        $DatabaseCredentials,
+
+        [Parameter()]
         [ValidateSet("Present", "Absent")]
         [System.String]
         $Ensure = "Present",
@@ -105,9 +121,21 @@ function Set-TargetResource
             else
             {
                 Write-Verbose -Message "Enabling SPSessionState"
+                if ($params.UseSQLAuthentication -eq $true)
+                {
+                    Write-Verbose -Message "Using SQL authentication to create service application as `$useSQLAuthentication is set to $($params.useSQLAuthentication)."
+                    $databaseCredentialsParam = @{
+                        DatabaseCredentials = $params.DatabaseCredentials
+                    }
+                }
+                else
+                {
+                    $databaseCredentialsParam = ""
+                }
                 Enable-SPSessionStateService -DatabaseName $params.DatabaseName `
                     -DatabaseServer $params.DatabaseServer `
-                    -SessionTimeout $params.SessionTimeout
+                    -SessionTimeout $params.SessionTimeout `
+                    @databaseCredentialsParam
             }
         }
     }
@@ -143,6 +171,14 @@ function Test-TargetResource
         [Parameter(Mandatory = $true)]
         [System.String]
         $DatabaseServer,
+
+        [Parameter()]
+        [System.Boolean]
+        $UseSQLAuthentication,
+
+        [Parameter()]
+        [System.Management.Automation.PSCredential]
+        $DatabaseCredentials,
 
         [Parameter()]
         [ValidateSet("Present", "Absent")]