+++ tests/Unit/SharePointDsc/SharePointDsc.SPSecureStoreServiceApp.Tests.ps1
@@ -292,13 +292,12 @@ try
 
             Context -Name "When specific windows credentials are to be used for the database" -Fixture {
                 $testParams = @{
-                    Name                       = "Secure Store Service Application"
-                    ApplicationPool            = "SharePoint Search Services"
-                    AuditingEnabled            = $false
-                    DatabaseName               = "SP_ManagedMetadata"
-                    DatabaseCredentials        = $mockCredential
-                    DatabaseAuthenticationType = "Windows"
-                    Ensure                     = "Present"
+                    Name                = "Secure Store Service Application"
+                    ApplicationPool     = "SharePoint Search Services"
+                    AuditingEnabled     = $false
+                    DatabaseName        = "SP_ManagedMetadata"
+                    DatabaseCredentials = $mockCredential
+                    Ensure              = "Present"
                 }
 
                 Mock -CommandName Get-SPServiceApplication -MockWith {
@@ -310,27 +309,16 @@ try
                     Assert-MockCalled New-SPSecureStoreServiceApplication
                 }
 
-                It "Should throw an exception if database authentication type is not specified" {
-                    $testParams.Remove("DatabaseAuthenticationType")
-                    { Set-TargetResource @testParams } | Should Throw
-                }
-
-                It "Should throw an exception if the credentials aren't provided and the authentication type is set" {
-                    $testParams.Add("DatabaseAuthenticationType", "Windows")
-                    $testParams.Remove("DatabaseCredentials")
-                    { Set-TargetResource @testParams } | Should Throw
-                }
             }
 
             Context -Name "When specific SQL credentials are to be used for the database" -Fixture {
                 $testParams = @{
-                    Name                       = "Secure Store Service Application"
-                    ApplicationPool            = "SharePoint Search Services"
-                    AuditingEnabled            = $false
-                    DatabaseName               = "SP_ManagedMetadata"
-                    DatabaseCredentials        = $mockCredential
-                    DatabaseAuthenticationType = "SQL"
-                    Ensure                     = "Present"
+                    Name                = "Secure Store Service Application"
+                    ApplicationPool     = "SharePoint Search Services"
+                    AuditingEnabled     = $false
+                    DatabaseName        = "SP_ManagedMetadata"
+                    DatabaseCredentials = $mockCredential
+                    Ensure              = "Present"
                 }
 
                 Mock -CommandName Get-SPServiceApplication -MockWith { return $null }
@@ -340,16 +328,6 @@ try
                     Assert-MockCalled New-SPSecureStoreServiceApplication
                 }
 
-                It "Should throw an exception if database authentication type is not specified" {
-                    $testParams.Remove("DatabaseAuthenticationType")
-                    { Set-TargetResource @testParams } | Should Throw
-                }
-
-                It "Should throw an exception if the credentials aren't provided and the authentication type is set" {
-                    $testParams.Add("DatabaseAuthenticationType", "Windows")
-                    $testParams.Remove("DatabaseCredentials")
-                    { Set-TargetResource @testParams } | Should Throw
-                }
             }
 
             Context -Name "When the service app exists but it shouldn't" -Fixture {